Changed name of interfaces BoardImpl, PlayerImpl and ActionImpl, to Board, Player and Action. Class Player changed to PlayerImpl.
