+++ src/main/java/inf112/isolasjonsteamet/roborally/actions/Destruction.java
@@ -1,15 +1,15 @@
 package inf112.isolasjonsteamet.roborally.actions;
 
-import inf112.isolasjonsteamet.roborally.board.BoardImpl;
-import inf112.isolasjonsteamet.roborally.players.PlayerImpl;
+import inf112.isolasjonsteamet.roborally.board.Board;
+import inf112.isolasjonsteamet.roborally.players.Player;
 
 /**
  * Is an Action that let's the player destroy things on the board.
  */
-public class Destruction implements ActionImpl {
+public class Destruction implements Action {
 
 	@Override
-	public void perform(BoardImpl board, PlayerImpl player) {
+	public void perform(Board board, Player player) {
 
 	}
 }
