+++ src/main/java/inf112/isolasjonsteamet/roborally/actions/KillRobot.java
@@ -1,16 +1,16 @@
 package inf112.isolasjonsteamet.roborally.actions;
 
-import inf112.isolasjonsteamet.roborally.board.BoardImpl;
-import inf112.isolasjonsteamet.roborally.players.PlayerImpl;
+import inf112.isolasjonsteamet.roborally.board.Board;
+import inf112.isolasjonsteamet.roborally.players.Player;
 
 /**
  * Kill robot is an action that let's a player kill a robot or the main master robot.
  *
  */
-public class KillRobot implements ActionImpl {
+public class KillRobot implements Action {
 
 	@Override
-	public void perform(BoardImpl board, PlayerImpl player) {
+	public void perform(Board board, Player player) {
 		//TODO
 		//player.kill();
