+++ src/main/java/inf112/isolasjonsteamet/roborally/actions/PowerDown.java
@@ -1,16 +1,16 @@
 package inf112.isolasjonsteamet.roborally.actions;
 
-import inf112.isolasjonsteamet.roborally.board.BoardImpl;
-import inf112.isolasjonsteamet.roborally.players.PlayerImpl;
+import inf112.isolasjonsteamet.roborally.board.Board;
+import inf112.isolasjonsteamet.roborally.players.Player;
 
 /**
  * Power Down is an Action that let's the player take a break with their robot while the game is
  * still going.
  */
-public class PowerDown implements ActionImpl {
+public class PowerDown implements Action {
 
 	@Override
-	public void perform(BoardImpl board, PlayerImpl player) {
+	public void perform(Board board, Player player) {
 		//TODO
 		//player.powerDown();
