+++ src/main/java/inf112/isolasjonsteamet/roborally/actions/RepairPlayer.java
@@ -1,15 +1,15 @@
 package inf112.isolasjonsteamet.roborally.actions;
 
-import inf112.isolasjonsteamet.roborally.board.BoardImpl;
-import inf112.isolasjonsteamet.roborally.players.PlayerImpl;
+import inf112.isolasjonsteamet.roborally.board.Board;
+import inf112.isolasjonsteamet.roborally.players.Player;
 
 /**
  * Repair player is an action that repairs a player x much according to cards they recieve.
  */
-public class RepairPlayer implements ActionImpl {
+public class RepairPlayer implements Action {
 
 	@Override
-	public void perform(BoardImpl board, PlayerImpl player) {
+	public void perform(Board board, Player player) {
 		//TODO
 		//player.repair();
