+++ src/main/java/inf112/isolasjonsteamet/roborally/actions/Reverse.java
@@ -1,13 +1,12 @@
 package inf112.isolasjonsteamet.roborally.actions;
 
-import inf112.isolasjonsteamet.roborally.board.BoardImpl;
+import inf112.isolasjonsteamet.roborally.board.Board;
 import inf112.isolasjonsteamet.roborally.players.Player;
-import inf112.isolasjonsteamet.roborally.players.PlayerImpl;
 
 /**
  * An Action that let's the robot go backwards.
  */
-public class Reverse implements ActionImpl {
+public class Reverse implements Action {
 
 	public Reverse() {}
 
@@ -15,7 +14,7 @@ public class Reverse implements ActionImpl {
 	 * Perfom an Action on the board.
 	 */
 	@Override
-	public void perform(BoardImpl board, PlayerImpl player) {
+	public void perform(Board board, Player player) {
 		player.setMovDir(player.getDir().getOpposingDir());
 		new Move().perform(board, player);
 	}
