+++ src/main/java/inf112/isolasjonsteamet/roborally/actions/Uturn.java
@@ -1,15 +1,15 @@
 package inf112.isolasjonsteamet.roborally.actions;
 
-import inf112.isolasjonsteamet.roborally.board.BoardImpl;
-import inf112.isolasjonsteamet.roborally.players.PlayerImpl;
+import inf112.isolasjonsteamet.roborally.board.Board;
+import inf112.isolasjonsteamet.roborally.players.Player;
 
 /**
  * An action that let's the player make a turn in the opposite direction.
  */
-public class Uturn implements ActionImpl {
+public class Uturn implements Action {
 
 	@Override
-	public void perform(BoardImpl board, PlayerImpl player) {
+	public void perform(Board board, Player player) {
 
 	}
 }
