+++ src/main/java/inf112/isolasjonsteamet/roborally/board/Board.java
@@ -1,24 +1,24 @@
 package inf112.isolasjonsteamet.roborally.board;
 
 import com.badlogic.gdx.math.Vector2;
-import inf112.isolasjonsteamet.roborally.players.PlayerImpl;
+import inf112.isolasjonsteamet.roborally.players.Player;
 import inf112.isolasjonsteamet.roborally.tiles.TileType;
 import java.util.List;
 import javax.annotation.Nullable;
 
-public interface BoardImpl {
+public interface Board {
 
 	/**
 	 * Get a list of the Players on the Board.
 	 * @return playerList
 	 */
-	List<PlayerImpl> getPlayers();
+	List<Player> getPlayers();
 
 	@Nullable
 	/**
 	 * Get player at a given osition.
 	 */
-	PlayerImpl getPlayerAt(Vector2 pos);
+	Player getPlayerAt(Vector2 pos);
 
 	/**
 	 * Get tiles at a given position.
